detectors/yolo/requirements.txt
torch==2.5.1
transformers>=4.47.0
pillow==10.4.0
requests==2.32.3
timm==1.0.12
opencv-python-headless==4.10.0.84
clarifai==11.4.2
clarifai-protocol>=0.0.23
ultralytics